Rename and move trace mapping component file
